pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.8",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
]
